+++ src/main/java/games/cubi/raycastedEntityOcclusion/SnapshotListener.java
@@ -1,5 +1,6 @@
 package games.cubi.raycastedEntityOcclusion;
 
+import games.cubi.raycastedEntityOcclusion.Snapshot.ChunkSnapshotManager;
 import org.bukkit.Material;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
